refactoring: fix decorators order, add "id" validation in projects.

+++ src/projects/projects.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
-import { Model } from 'mongoose';
+import mongoose, { Model } from 'mongoose';
 import { Project } from './model/project.schema';
 
 @Injectable()
@@ -22,6 +22,9 @@ export class ProjectsService {
   }
 
   async findById(id: string): Promise<Project> {
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      return null;
+    }
     return this.projectModel.findById(id);
   }
 
@@ -30,6 +33,9 @@ export class ProjectsService {
     name: string,
     description: string,
   ): Promise<Project> {
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      return null;
+    }
     return this.projectModel.findByIdAndUpdate(
       id,
       { name: name, description: description },
@@ -38,6 +44,9 @@ export class ProjectsService {
   }
 
   async delete(id: string): Promise<Project> {
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      return null;
+    }
     return this.projectModel.findByIdAndDelete(id);
   }
 }
